Simple series test
Checks for series header and container
